+++ sql/emaj--devel.sql
@@ -5893,6 +5893,52 @@ $_drop_group$
   END;
 $_drop_group$;
 
+CREATE OR REPLACE FUNCTION emaj.emaj_forget_group(p_groupName TEXT)
+RETURNS INT LANGUAGE plpgsql AS
+$emaj_forget_group$
+-- This function deletes all traces of a dropped group from emaj_group_hist and emaj_log_sessions tables.
+-- Input: group name
+-- Output: number of deleted rows
+  DECLARE
+    v_nbDeletedSession       INT = 0;
+    v_nbDeletedHistory       INT = 0;
+  BEGIN
+-- Insert a BEGIN event into the history.
+    INSERT INTO emaj.emaj_hist (hist_function, hist_event, hist_object)
+      VALUES ('FORGET_GROUP', 'BEGIN', p_groupName);
+-- Check that the group is not recorded in emaj_group table anymore
+    IF EXISTS
+         (SELECT 0
+            FROM emaj.emaj_group
+            WHERE group_name = p_groupName
+         ) THEN
+      RAISE EXCEPTION 'emaj_forget_group: The group "%" still exists.', p_groupName;
+    END IF;
+-- OK
+-- Delete rows from emaj_log_session.
+    DELETE FROM emaj.emaj_log_session
+      WHERE lses_group = p_groupName;
+    GET DIAGNOSTICS v_nbDeletedSession = ROW_COUNT;
+-- Delete rows from emaj_group_hist.
+    DELETE FROM emaj.emaj_group_hist
+      WHERE grph_group = p_groupName;
+    GET DIAGNOSTICS v_nbDeletedHistory = ROW_COUNT;
+-- Warn if the group has not been found in any history table.
+    IF v_nbDeletedSession + v_nbDeletedHistory = 0 THEN
+      RAISE WARNING 'emaj_forget_group: the tables group "%" has not been found in history tables', p_groupName;
+    END IF;
+-- Insert a END event into the history.
+    INSERT INTO emaj.emaj_hist (hist_function, hist_event, hist_object, hist_wording)
+      VALUES ('FORGET_GROUP', 'END', p_groupName,
+              v_nbDeletedSession || ' rows deleted from emaj_log_session and ' ||
+              v_nbDeletedHistory || ' rows deleted from emaj_group_hist');
+--
+    RETURN v_nbDeletedSession + v_nbDeletedHistory;
+  END;
+$emaj_forget_group$;
+COMMENT ON FUNCTION emaj.emaj_forget_group(TEXT) IS
+$$Removes traces of a dropped group from histories.$$;
+
 CREATE OR REPLACE FUNCTION emaj.emaj_export_groups_configuration(p_groups TEXT[] DEFAULT NULL)
 RETURNS JSON LANGUAGE plpgsql AS
 $emaj_export_groups_configuration$